+++ changelog.md
@@ -45,6 +45,9 @@
 - Added `sugar.capture` for capturing some local loop variables when creating a closure.
   This is an enhanced version of `closureScope`.
 
+- Added `count`, `countIt`, `lengthWhile`, `takeWhile`, `dropWhile`, `nest`, `zipwith`,
+  and `findFirst` to `sequtils` module to enrich its functionalities.
+
 ## Library changes
 
 - `asyncdispatch.drain` now properly takes into account `selector.hasPendingOperations`